+++ ethers-providers/src/provider.rs
@@ -1,5 +1,5 @@
 use crate::{
-    ens, erc, maybe,
+    call_raw, ens, erc, maybe,
     pubsub::{PubsubClient, SubscriptionStream},
     stream::{FilterWatcher, DEFAULT_POLL_INTERVAL},
     FromErr, Http as HttpProvider, JsonRpcClient, JsonRpcClientWrapper, LogQuery, MockProvider,
@@ -229,6 +229,10 @@ impl<P: JsonRpcClient> Provider<P> {
                 self.request("eth_getBlockByNumber", [num, include_txs]).await?
             }
         })
+    }
+
+    pub fn call_raw<'a>(&'a self, tx: &'a TypedTransaction) -> call_raw::Call<'a, P> {
+        call_raw::Call::new(self, tx)
     }
 }
 
@@ -1161,275 +1165,6 @@ impl<P: JsonRpcClient> Middleware for Provider<P> {
     }
 }
 
-use crate::PinBoxFut;
-use serde::ser::SerializeTuple;
-use std::{
-    collections::HashMap,
-    future::Future,
-    pin::Pin,
-    task::{Context, Poll},
-};
-
-impl<P: JsonRpcClient> Provider<P> {
-    pub fn call_builder<'a>(&'a self, tx: &'a TypedTransaction) -> Caller<'a, P> {
-        Caller::Build(CallBuilder::new(self, tx))
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CallInput<'a> {
-    tx: &'a TypedTransaction,
-    block: Option<BlockId>,
-    spoof: Option<&'a spoof::State>,
-}
-impl<'a> CallInput<'a> {
-    pub fn new(tx: &'a TypedTransaction) -> Self {
-        Self { tx, block: None, spoof: None }
-    }
-}
-
-impl<'a> Serialize for CallInput<'a> {
-    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
-    where
-        S: serde::ser::Serializer,
-    {
-        let len = 2 + self.spoof.is_some() as usize;
-
-        let mut tup = serializer.serialize_tuple(len)?;
-        tup.serialize_element(self.tx)?;
-
-        let block = self.block.unwrap_or_else(|| BlockNumber::Latest.into());
-        tup.serialize_element(&block)?;
-
-        if let Some(spoof) = self.spoof {
-            tup.serialize_element(spoof)?;
-        }
-        tup.end()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct CallBuilder<'a, P> {
-    provider: &'a Provider<P>,
-    input: CallInput<'a>,
-}
-
-impl<'a, P> CallBuilder<'a, P> {
-    fn new(provider: &'a Provider<P>, tx: &'a TypedTransaction) -> Self {
-        Self { provider, input: CallInput::new(tx) }
-    }
-}
-impl<'a, P: JsonRpcClient> CallBuilder<'a, P> {
-    fn execute(&self) -> impl Future<Output = Result<Bytes, ProviderError>> + 'a {
-        self.provider.request("eth_call", utils::serialize(&self.input))
-    }
-}
-
-pub enum Caller<'a, P> {
-    Build(CallBuilder<'a, P>),
-    Wait(PinBoxFut<'a, Bytes>),
-}
-
-impl<'a, P> Caller<'a, P> {
-    pub fn map<F>(self, f: F) -> Self
-    where
-        F: FnOnce(&mut CallBuilder<'a, P>),
-    {
-        match self {
-            Self::Build(mut call) => {
-                f(&mut call);
-                Self::Build(call)
-            }
-            wait => wait,
-        }
-    }
-    pub fn block(self, id: BlockId) -> Self {
-        self.map(|mut call| call.input.block = Some(id))
-    }
-    pub fn spoof(self, state: &'a spoof::State) -> Self {
-        self.map(|mut call| call.input.spoof = Some(state))
-    }
-    pub fn unwrap_builder(self) -> CallBuilder<'a, P> {
-        match self {
-            Self::Build(b) => b,
-            _ => panic!("Caller::unwrap_builder on a Wait value"),
-        }
-    }
-}
-
-impl<'a, P: JsonRpcClient> Future for Caller<'a, P> {
-    type Output = Result<Bytes, ProviderError>;
-
-    fn poll(self: Pin<&mut Self>, ctx: &mut Context) -> Poll<Self::Output> {
-        let pin = self.get_mut();
-        loop {
-            match pin {
-                Caller::Build(ref call) => {
-                    let fut = Box::pin(call.execute());
-                    *pin = Caller::Wait(fut);
-                }
-                Caller::Wait(ref mut fut) => return fut.as_mut().poll(ctx),
-            }
-        }
-    }
-}
-
-/// Provides types and methods for "spoofing" state overrides for eth_call
-pub mod spoof {
-    use super::*;
-
-    #[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-    pub struct Account {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub nonce: Option<U64>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub balance: Option<U256>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub code: Option<Bytes>,
-        #[serde(flatten, skip_serializing_if = "Option::is_none")]
-        pub storage: Option<Storage>,
-    }
-
-    impl Account {
-        pub fn nonce(&mut self, nonce: U64) -> &mut Self {
-            self.nonce = Some(nonce);
-            self
-        }
-        pub fn balance(&mut self, bal: U256) -> &mut Self {
-            self.balance = Some(bal);
-            self
-        }
-        pub fn code(&mut self, code: Bytes) -> &mut Self {
-            self.code = Some(code);
-            self
-        }
-        pub fn store(&mut self, key: H256, val: U256) -> &mut Self {
-            self.storage.get_or_insert_with(Default::default).insert(key, val);
-            self
-        }
-    }
-
-    #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-    pub enum Storage {
-        #[serde(rename = "stateDiff")]
-        Diff(HashMap<H256, U256>),
-        #[serde(rename = "state")]
-        Replace(HashMap<H256, U256>),
-    }
-
-    impl Default for Storage {
-        fn default() -> Self {
-            Self::Diff(Default::default())
-        }
-    }
-    impl std::ops::Deref for Storage {
-        type Target = HashMap<H256, U256>;
-        fn deref(&self) -> &Self::Target {
-            match self {
-                Self::Diff(map) => map,
-                Self::Replace(map) => map,
-            }
-        }
-    }
-    impl std::ops::DerefMut for Storage {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            match self {
-                Self::Diff(map) => map,
-                Self::Replace(map) => map,
-            }
-        }
-    }
-
-    #[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-    #[serde(transparent)]
-    pub struct State(#[serde(skip_serializing_if = "HashMap::is_empty")] HashMap<Address, Account>);
-
-    impl State {
-        pub fn account(&mut self, adr: Address) -> &mut Account {
-            self.0.entry(adr).or_default()
-        }
-    }
-
-    /// # Example
-    /// ```no_run
-    /// # use ethers_core::{
-    /// #     types::{Address, TransactionRequest, H256},
-    /// #     utils::{parse_ether},
-    /// # };
-    /// # use ethers_providers::{Provider, Http, Middleware, spoof};
-    /// # use std::convert::TryFrom;
-    /// #
-    /// # #[tokio::main(flavor = "current_thread")]
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let provider = Provider::<Http>::try_from("http://127.0.0.1:8545").unwrap();
-    ///
-    /// let adr1: Address = "0x6fC21092DA55B392b045eD78F4732bff3C580e2c".parse().unwrap();
-    /// let adr2: Address = "0x295a70b2de5e3953354a6a8344e616ed314d7251".parse().unwrap();
-    ///
-    /// let tx = Default::default();
-    ///
-    /// let mut state = spoof::state();
-    /// state.account(adr1).store(H256::default(), 1.into()).nonce(2.into());
-    /// provider.call_builder(&tx).spoof(&state).await.unwrap();
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn state() -> State {
-        State::new()
-    }
-
-    /// # Example
-    /// ```no_run
-    /// # use ethers_core::{
-    /// #     types::{Address, TransactionRequest, H256},
-    /// #     utils::{parse_ether},
-    /// # };
-    /// # use ethers_providers::{Provider, Http, Middleware, spoof};
-    /// # use std::convert::TryFrom;
-    /// # #[tokio::main(flavor = "current_thread")]
-    /// #
-    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let provider = Provider::<Http>::try_from("http://127.0.0.1:8545").unwrap();
-    ///
-    /// let adr1: Address = "0x6fC21092DA55B392b045eD78F4732bff3C580e2c".parse()?;
-    /// let adr2: Address = "0x295a70b2de5e3953354a6a8344e616ed314d7251".parse()?;
-    /// let pay_amt = parse_ether(1u64)?;
-    ///
-    /// // Not enough ether to pay for the transaction
-    /// let tx = TransactionRequest::pay(adr2, pay_amt).from(adr1);
-    /// let tx = tx.into();
-    ///
-    /// // override the sender's balance for the call
-    /// let mut state = spoof::balance(adr1, pay_amt * 2);
-    /// provider.call_builder(&tx).spoof(&state).await?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn balance(adr: Address, bal: U256) -> State {
-        let mut state = State::new();
-        state.account(adr).balance(bal);
-        state
-    }
-
-    pub fn nonce(adr: Address, nonce: U64) -> State {
-        let mut state = State::new();
-        state.account(adr).nonce(nonce);
-        state
-    }
-
-    pub fn code(adr: Address, code: Bytes) -> State {
-        let mut state = State::new();
-        state.account(adr).code(code);
-        state
-    }
-
-    pub fn storage(adr: Address, key: H256, val: U256) -> State {
-        let mut state = State::new();
-        state.account(adr).store(key, val);
-        state
-    }
-}
-
 impl<P: JsonRpcClient> Provider<P> {
     async fn query_resolver<T: Detokenize>(
         &self,
@@ -2190,67 +1925,5 @@ mod tests {
         assert_eq!(tx.gas(), Some(&gas));
         assert_eq!(tx.gas_price(), Some(gas_price));
         assert!(tx.access_list().is_none());
-    }
-
-    #[derive(Debug, Deserialize)]
-    struct CallInputOwned(
-        TypedTransaction,
-        Option<BlockId>,
-        #[serde(default)] Option<spoof::State>,
-    );
-    impl<'a> From<&'a CallInputOwned> for CallInput<'a> {
-        fn from(src: &'a CallInputOwned) -> Self {
-            Self { tx: &src.0, block: src.1, spoof: src.2.as_ref() }
-        }
-    }
-
-    // Tests "roundtrip" serialization of calls: deserialize(serialize(call)) == call
-    fn test_encode<'a, P>(call: Caller<'a, P>) {
-        let input = call.unwrap_builder().input;
-        let ser = utils::serialize(&input).to_string();
-        let de: CallInputOwned = serde_json::from_str(&ser).unwrap();
-        let de = CallInput::from(&de);
-
-        assert_eq!(input.tx, de.tx);
-        assert_eq!(input.spoof, de.spoof);
-
-        let block = input.block.or_else(|| Some(BlockNumber::Latest.into()));
-        assert_eq!(block, de.block);
-    }
-
-    #[test]
-    fn test_serialize_call() {
-        let adr1: Address = "0x6fC21092DA55B392b045eD78F4732bff3C580e2c".parse().unwrap();
-        let adr2: Address = "0x295a70b2de5e3953354a6a8344e616ed314d7251".parse().unwrap();
-        let h1 = utils::keccak256("foo").into();
-        let h2 = utils::keccak256("bar").into();
-
-        let tx = TypedTransaction::default();
-        let (provider, _) = Provider::mocked();
-
-        let call = provider.call_builder(&tx);
-        test_encode(call);
-
-        let mut state = spoof::state();
-        state
-            .account(adr1)
-            .nonce(1.into())
-            .balance(2.into())
-            .store(h1, 534.into())
-            .store(h2, 35.into());
-        let call = provider.call_builder(&tx).block(100.into()).spoof(&state);
-        test_encode(call);
-
-        let mut state = spoof::state();
-        state.account(adr1).nonce(1.into());
-        state.account(adr2).nonce(7.into());
-        let call = provider.call_builder(&tx).spoof(&state);
-        test_encode(call);
-
-        // State override with an empty acccount should be encoded as "0xab..": {}
-        let mut state = spoof::state();
-        state.account(adr1);
-        let call = provider.call_builder(&tx).spoof(&state);
-        test_encode(call);
     }
 }
